refactor: Implement modular architecture for Dockit Elements
- Restructured codebase into a modular format for improved clarity, maintainability, and scalability.
- Created core modules: constants, types, style-registry, element-utils, element, component, and root.
- Introduced a dedicated elements file for standard HTML element factories.
- Ensured backward compatibility with existing API.
- Updated documentation to reflect new architecture and usage guidelines.
- Enhanced testing plan to cover new modular structure.
