+++ src/core/constants.ts
@@ -1,0 +1,1 @@
+export const CLASS_PREFIX = 'dockit-'; // Constant for class name prefix
